requirements.txt
asgiref==3.2.7
autopep8==1.5.4
cachetools==4.1.1
certifi==2020.4.5.1
chardet==3.0.4
coreapi==2.3.3
coreschema==0.0.4
Django==3.0.7
django-cors-headers==3.5.0
djangorestframework==3.11.0
djangorestframework-simplejwt==4.4.0
drf-yasg==1.17.1
facebook-sdk==3.1.0
Faker==4.1.0
future==0.18.2
google-api-core==1.22.2
google-api-python-client==1.12.1
google-auth==1.21.2
google-auth-httplib2==0.0.4
googleapis-common-protos==1.52.0
gunicorn==20.0.4
httplib2==0.18.1
idna==2.9
inflection==0.4.0
itypes==1.2.0
Jinja2==2.11.2
MarkupSafe==1.1.1
oauthlib==3.1.0
packaging==20.4
protobuf==3.13.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycodestyle==2.6.0
PyJWT==1.7.1
pyparsing==2.4.7
python-dateutil==2.8.1
python-twitter==3.5
pytz==2020.1
requests==2.23.0
requests-oauthlib==1.3.0
rsa==4.6
ruamel.yaml==0.16.10
ruamel.yaml.clib==0.2.0
six==1.15.0
sqlparse==0.3.1
text-unidecode==1.3
toml==0.10.1
uritemplate==3.0.1
urllib3==1.25.9